% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ratio.R
\name{calculate_ppi_ratio}
\alias{calculate_ppi_ratio}
\title{Calculate ratio for PPI}
\usage{
calculate_ppi_ratio(ppi, input_genes)
}
\arguments{
\item{ppi}{A df of ppi annotations}

\item{input_genes}{A df of input genes}
}
\value{
A dataframe with the fiveP scores
}
\description{
Calculate ratio for PPI
}
