% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_gene_list.R
\name{fetch_hgnc_gene_list}
\alias{fetch_hgnc_gene_list}
\title{Fetch Data}
\usage{
fetch_hgnc_gene_list()
}
\value{
A dataframe with data from Database.
}
\description{
This function fetches data from Database for a given list of genes.
}
