% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_families.R
\name{fetch_protein_families}
\alias{fetch_protein_families}
\title{Fetch Data}
\usage{
fetch_protein_families(
  protein_coding_genes,
  uniprot_input_gene_symbol_results_cleaned
)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{uniprot_input_gene_symbol_results_cleaned}{The df of uniprot results of input genes.}
}
\value{
A dataframe with data from Database.
}
\description{
This function fetches data from Database for a given list of genes.
}
