% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{fetch_uniprot}
\alias{fetch_uniprot}
\title{Fetch Data}
\usage{
fetch_uniprot(protein_coding_genes, input_genes)
}
\arguments{
\item{protein_coding_genes}{The df of all protein coding genes.}

\item{input_genes}{The df of input genes.}
}
\value{
A dataframe with data from Database.
}
\description{
This function fetches data from Database for a given list of genes.
}
