% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_fiveP}
\alias{get_fiveP}
\title{Fetch and process data}
\usage{
get_fiveP(input_genes)
}
\arguments{
\item{input_genes}{A vector of gene HGNC IDs}
}
\value{
A dataframe with the fiveP scores
}
\description{
Fetch and process data
}
